Build system fix attempt
