+++ vhdl_parser_glue.cpp
@@ -68,32 +68,3 @@ VhdlParseTreeNode *YaVHDL::Parser::VhdlParserParseFile(
 
     return parse_output;
 }
-
-#ifdef VHDL_PARSER_DEMO_MODE
-
-#include <iostream>
-using namespace std;
-
-int main(int argc, char **argv) {
-    if (argc < 2) {
-        cout << "Usage: " << argv[0] << " file.vhd\n";
-        return -1;
-    }
-
-    int ret = 0;
-
-    std::string errors = std::string();
-    VhdlParseTreeNode *parse_output = VhdlParserParseFile(argv[1], errors);
-    if (parse_output) {
-        parse_output->debug_print();
-    } else {
-        ret = 1;
-        cout << errors;
-    }
-    cout << "\n";
-    delete parse_output;
-
-    return ret;
-}
-
-#endif
